droid: Only kill processes on logout if TTY is /dev/pts/0
When logout from SSHD, those processes should not be killed

+++ droid/sshd/default.nix
@@ -38,6 +38,6 @@ in
   ];
 
   hm.programs.zsh.logoutExtra = ''
-    pkill sshd
+    [ "$TTY" = /dev/pts/0 ] && pkill sshd
   '';
 }
